Enterprise text_rag_raw_langchain/requirements.txt
openai
python-dotenv
tiktoken
numpy
faiss-cpu
tqdm
pypdf
fastapi
uvicorn[standard]
pydantic
langchain
langchain-community 
langchain-openai
